Fixed fitter:
Since measurements are first updated and then predicted, the fitter was
returning the measurements moved by one. So I moved the extra final filter step
to before the loop over the measurements. That way the filter is at the first
measurement when the loop begins and there is no shift between prediction and
measurement.